+++ sway-lsp/src/lib.rs
@@ -6,18 +6,25 @@ mod capabilities;
 pub mod config;
 mod core;
 pub mod error;
+pub mod global_state;
+mod handlers {
+    pub mod notification;
+    pub mod request;
+}
+pub mod lsp_ext;
 pub mod server;
 mod traverse;
 pub mod utils;
-use server::Backend;
+
+use global_state::GlobalState;
 
 pub async fn start() {
     let stdin = tokio::io::stdin();
     let stdout = tokio::io::stdout();
 
-    let (service, socket) = LspService::build(Backend::new)
-        .custom_method("sway/show_ast", Backend::show_ast)
-        .custom_method("textDocument/inlayHint", Backend::inlay_hints)
+    let (service, socket) = LspService::build(GlobalState::new)
+        .custom_method("sway/show_ast", handlers::request::handle_show_ast)
+        .custom_method("textDocument/inlayHint", handlers::request::handle_inlay_hints)
         .finish();
     Server::new(stdin, stdout, socket).serve(service).await;
 }
