+++ sway-lsp/src/lsp_ext.rs
@@ -1,0 +1,12 @@
+//! sway-lsp extensions to the LSP.
+
+use lsp_types::{request::Request, TextDocumentIdentifier, Url};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ShowAstParams {
+    pub text_document: TextDocumentIdentifier,
+    pub ast_kind: String,
+    pub save_path: Url,
+}
